+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/source/ServiceInstance.java
@@ -71,10 +71,6 @@ public class ServiceInstance extends Source {
     private boolean status;
     @Getter
     @Setter
-    @Deprecated
-    private int responseCode;
-    @Getter
-    @Setter
     private int httpResponseStatusCode;
     @Getter
     @Setter